refactor: Question 컴포넌트화
