+++ src/pages/home/index.tsx
@@ -2,9 +2,7 @@ import PopularHot from '@/components/PoularHot';
 
 function Home() {
   return (
-    <>
-      <PopularHot></PopularHot>
-    </>
+    <PopularHot/>
   );
 }
